major refactor, new features
Bischop is WIP, building of fat jar not yet working

+++ MetaChessApp/src/userinterface/specific/GTIcon.java
@@ -5,7 +5,7 @@ import userinterface.generic.GUI1Tile;
 import userinterface.generic.GUITile;
 
 public class GTIcon extends GUI1Tile{
-	public GTIcon(int color, GUITile container, int i, int j,String name) {
+	public GTIcon(float color, GUITile container, int i, int j,String name) {
 		super(color, container, i, j);
 		grid=name;
